refactor: extract LocomotionBaseEnv from G1/Go1/Go2 base environments
Closes #133

- Create `locomotion/common/` module with shared base class, configs, and utilities
- `LocomotionBaseEnv` consolidates 7 identical methods from 3 robot base envs
- `LocomotionBaseCfg` + `ControlConfigBase` provide shared config inheritance
- `DomainRandConfig` replaces `Domain_Rand` (PascalCase naming)
- `Commands` + `sample_velocity_commands` extracted to common module
- Line count reductions: G1 52%, Go1 62%, Go2 50%
- Fix pre-existing mypy type-arg error in xml_utils.py
