+++ src/unilab/envs/locomotion/common/__init__.py
@@ -1,0 +1,13 @@
+from .base import ControlConfigBase, LocomotionBaseCfg, LocomotionBaseEnv, Sensor
+from .commands import Commands, sample_velocity_commands
+from .domain_rand import DomainRandConfig
+
+__all__ = [
+    "Commands",
+    "ControlConfigBase",
+    "DomainRandConfig",
+    "LocomotionBaseCfg",
+    "LocomotionBaseEnv",
+    "Sensor",
+    "sample_velocity_commands",
+]
