+++ src/unilab/envs/locomotion/go1/base.py
@@ -2,13 +2,11 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 
-import gymnasium as gym
-import numpy as np
-
-from unilab.base.backend import SimBackend
-from unilab.base.base import EnvCfg
-from unilab.base.dtype_config import get_global_dtype
-from unilab.base.np_env import NpEnv, NpEnvState
+from unilab.envs.locomotion.common.base import (
+    ControlConfigBase,
+    LocomotionBaseCfg,
+    LocomotionBaseEnv,
+)
 
 
 @dataclass
@@ -22,11 +20,9 @@ class NoiseConfig:
 
 
 @dataclass
-class ControlConfig:
-    action_scale: float = 0.25
+class ControlConfig(ControlConfigBase):
     Kp: float = 35.0
     Kd: float = 0.5
-    simulate_action_latency: bool = False
 
 
 @dataclass
@@ -37,83 +33,13 @@ class Asset:
 
 
 @dataclass
-class Sensor:
-    local_linvel = "local_linvel"
-    gyro = "gyro"
-
-
-@dataclass
-class Domain_Rand:
-    # randomize_friction = True
-    # friction_range = [0.5, 1.25]
-    randomize_base_mass = False
-    added_mass_range = [-1.5, 1.5]
-
-    random_com = False
-    com_offset_x = [-0.05, 0.05]
-
-    push_robots = False
-    push_interval = 750  # step
-    max_force = [1, 1, 0.5]
-
-
-@dataclass
-class Go1BaseCfg(EnvCfg):
-    model_file: str = field(default=str(""))
+class Go1BaseCfg(LocomotionBaseCfg):
     noise_config: NoiseConfig = field(default_factory=NoiseConfig)
-    control_config: ControlConfig = field(default_factory=ControlConfig)
+    control_config: ControlConfig = field(default_factory=ControlConfig)  # type: ignore[assignment]
     asset: Asset = field(default_factory=Asset)
-    sensor: Sensor = field(default_factory=Sensor)
     sim_dt: float = 0.01
     ctrl_dt: float = 0.02
 
 
-class Go1BaseEnv(NpEnv):
+class Go1BaseEnv(LocomotionBaseEnv):
     _cfg: Go1BaseCfg
-
-    def __init__(self, cfg: Go1BaseCfg, backend: SimBackend, num_envs=1):
-        super().__init__(cfg, backend, num_envs)
-
-        self._init_action_space()
-        self._num_action = self._action_space.shape[0]
-        self._init_buffers()
-
-    def _init_action_space(self):
-        ctrl_range = self._backend.get_actuator_ctrl_range()
-        nu = self._backend.num_actuators
-        self._action_space = gym.spaces.Box(ctrl_range[:, 0], ctrl_range[:, 1], (nu,), dtype=float)  # type: ignore[assignment]
-
-    @property
-    def action_space(self) -> gym.spaces.Box:
-        return self._action_space  # type: ignore[no-any-return]
-
-    def _init_buffers(self):
-        dtype = get_global_dtype()
-        self.default_angles = np.zeros((self._num_action,), dtype=dtype)
-        self._init_qpos = np.array(self._backend.get_keyframe_qpos("home"), dtype=dtype)
-        self.default_angles = self._init_qpos[7:]
-        self._init_qvel = self._backend.get_init_qvel().astype(dtype)
-
-    def apply_action(self, actions: np.ndarray, state: NpEnvState) -> np.ndarray:
-        state.info["last_actions"] = state.info.get("current_actions", np.zeros_like(actions))
-        state.info["current_actions"] = actions
-        exec_actions = (
-            state.info["last_actions"]
-            if self._cfg.control_config.simulate_action_latency
-            else actions
-        )
-        return np.asarray(
-            exec_actions * self._cfg.control_config.action_scale + self.default_angles
-        )
-
-    def get_local_linvel(self) -> np.ndarray:
-        return np.asarray(self._backend.get_sensor_data(self._cfg.sensor.local_linvel))
-
-    def get_gyro(self) -> np.ndarray:
-        return np.asarray(self._backend.get_sensor_data(self._cfg.sensor.gyro))
-
-    def get_dof_pos(self) -> np.ndarray:
-        return np.asarray(self._backend.get_dof_pos())
-
-    def get_dof_vel(self) -> np.ndarray:
-        return np.asarray(self._backend.get_dof_vel())
