+++ src/unilab/utils/xml_utils.py
@@ -105,7 +105,7 @@ def _add_b_sensors(sensor_tag: ET.Element, valid_bnames: list[str], baselink_nam
         )
 
 
-def _write_temp_xml(tree: ET.ElementTree[ET.Element], model_file: str) -> str:
+def _write_temp_xml(tree: ET.ElementTree[ET.Element], model_file: str) -> str:  # type: ignore[type-arg]
     fd, output_path = tempfile.mkstemp(
         suffix=".xml", dir=os.path.dirname(os.path.abspath(model_file))
     )
